compile fixes

+++ api/oc_knx_client.c
@@ -570,7 +570,7 @@ oc_do_s_mode_read(size_t group_address)
   int iid = device->iid;
 
   PRINT("oc_do_s_mode_read : ga=%d ia=%d, iid=%d\n", 
-    group_address, sia_value, iid);
+    (int)group_address, sia_value, iid);
 
   if (group_address > 0) {
     oc_issue_s_mode(2, sia_value, group_address, iid, "r", 0, 0);
